site_et_services_web/requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-xml==2.0.0
PyJWT==2.7.0
PySimpleSOAP==1.16.2
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
